Prompt:
We are going to write a GitHub bot in Rust, that can help to make sure PRs are ok to merge, it should have support for multiple repositories. I should be stateful.
The idea is that it will merge the PR branch into master, in a separate branch called `automation/bot/try-merge` and `automation/bot/try`. It should also have support for commands via PR comments, for example `@bot try`. It should listen to GitHub webhooks concurrently, but handle commands serially, to avoid race conditions, if possible, we would like it to handle it all concurrently. The state should be stored in CockroachDB.